1.1

+++ lib/DocumentStoreOne.php
@@ -3,7 +3,7 @@ namespace eftec\DocumentStoreOne;
 
 /**
  * Class DocumentStoreOne
- * @version 1.0 2018-08-11
+ * @version 1.1 2018-08-12
  * @author Jorge Castro Castillo jcastro@eftec.cl
  * @license LGPLv3
  */
@@ -11,8 +11,8 @@ class DocumentStoreOne {
 
     /** @var string root folder of the database */
     var $database;
-    /** @var string schema (subfolder) of the database */
-    var $schema;
+    /** @var string collection (subfolder) of the database */
+    var $collection;
     /** @var int Maximium duration of the lock (in seconds). By default it's 2 minutes */
     var $maxLockTime=120;
     /** @var int Default number of retries. By default it tries 300x0.1sec=30 seconds */
@@ -20,19 +20,21 @@ class DocumentStoreOne {
     /** @var int Interval (in microseconds) between retries. 100000 means 0.1 seconds */
     var $intervalBetweenRetry=100000;
     /** @var string Default extension (with dot) of the document */
-    var $docExt=".json";
+    var $docExt=".dson";
+    /** @var null|string[] Indicates if it's locked manually. By default, every operation locks the document */
+    private $manualLock=null;
 
     /**
      * DocumentStoreOne constructor.
-     * @example $flatcon=new DocumentStoreOne(dirname(__FILE__)."/base",'schemaFolder');
+     * @example $flatcon=new DocumentStoreOne(dirname(__FILE__)."/base",'collectionFolder');
      * @param string $database root folder of the database
-     * @param string $schema schema (subfolder) of the database. If the schema is empty then it uses the root folder.
+     * @param string $collection collection (subfolder) of the database. If the collection is empty then it uses the root folder.
      * @throws \Exception
      */
-    public function __construct($database, $schema='')
+    public function __construct($database, $collection='')
     {
         $this->database = $database;
-        $this->schema = $schema;
+        $this->collection = $collection;
         if (!is_dir($this->getPath())) {
             throw new \Exception("Incorrect folder");
         }
@@ -52,6 +54,64 @@ class DocumentStoreOne {
     }
 
     /**
+     * @param string|string[] $Ids
+     * @param int $tries
+     * @return bool
+     */
+    public function lock($Ids, $tries=-1) {
+        if ($this->manualLock!=null) return false; // we can't lock because it's already locked. Unlock first.
+        if (is_array($Ids)) {
+            $v=-10;
+            $numItem=count($Ids);
+            for($i=0;$i<$numItem;$i++) {
+                if (!$this->lockFolder($Ids[$i], $tries)) {
+                    $v = $i-1; // it failed to lock the file.
+                    break;
+                }
+            }
+            if ($v!=-10) {
+                //something failed, rollback
+                for($i=0;$i<$v;$i++) {
+                    $this->unlockFolder($Ids[$i],true);
+                }
+                return false;
+            } else {
+                // the entire array is locked.
+                $this->manualLock=$Ids;
+                return true;
+            }
+        } else {
+            // we lock a single id.
+            if ($this->lockFolder($Ids,$tries)) {
+                $this->manualLock=array($Ids);
+                return true;
+            } else {
+                return false;
+            }
+        }
+    }
+
+    /**
+     * Manually unlock the document
+     * @return bool
+     */
+    public function unlock() {
+
+        if (is_array($this->manualLock)) {
+            $v=true;
+            var_dump($this->manualLock);
+            foreach($this->manualLock as $id) {
+                var_dump($id);
+                $v=$v && $this->unlockFolder($id,true);
+            }
+        } else {
+            $v=false; // is not locked.
+        }
+        $this->manualLock=null;
+        return $v;
+    }
+
+    /**
      * It gets the next sequence. If the sequence doesn't exist, it generates a new one with 1.
      * You could peek a sequence with get('genseq_<name>')
      * If the sequence is corrupt then it's resetted.
@@ -64,7 +124,7 @@ class DocumentStoreOne {
      */
     public function getNextSequence($name="seq",$tries=-1,$init=1,$interval=1,$reserveAdditional=0) {
         $id="genseq_".$name;
-        $file =$this->filename($id);
+        $file =$this->filename($id).".seq";
         if ($this->lockFolder($file,$tries)) {
             if (file_exists($file)) {
                 $read=@file_get_contents($file);
@@ -89,7 +149,7 @@ class DocumentStoreOne {
      * @return string
      */
     private function getPath() {
-        return $this->database."/".$this->schema;
+        return $this->database."/".$this->collection;
     }
 
     /**
@@ -158,45 +218,45 @@ class DocumentStoreOne {
     }
 
     /**
-     * Check a schema
-     * @param $schema
-     * @return bool It returns false if it's not a schema (a valid folder)
+     * Check a collection
+     * @param $collection
+     * @return bool It returns false if it's not a collection (a valid folder)
      */
-    public function isSchema($schema) {
-        $this->schema=$schema;
+    public function isCollection($collection) {
+        $this->collection=$collection;
         return is_dir($this->getPath());
     }
 
     /**
-     * Set the current schema
-     * @param $schema
+     * Set the current collection
+     * @param $collection
      * @return DocumentStoreOne
      */
-    public function schema($schema) {
-        $this->schema=$schema;
+    public function collection($collection) {
+        $this->collection=$collection;
         return $this;
     }
 
     /**
-     * Creates a schema
-     * @param $schema
+     * Creates a collection
+     * @param $collection
      * @return bool true if the operation is right, false if it fails.
      */
-    public function createSchema($schema) {
-        $oldSchema=$this->schema;
-        $this->schema=$schema;
+    public function createCollection($collection) {
+        $oldCollection=$this->collection;
+        $this->collection=$collection;
         $r=@mkdir($this->getPath());
-        $this->schema=$oldSchema;
+        $this->collection=$oldCollection;
         return $r;
     }
 
     /**
-     * List all the Ids in a schema.
+     * List all the Ids in a collection.
      * @param string $mask see http://php.net/manual/en/function.glob.php
      * @return array|false
      */
     public function select($mask="*") {
-        $list = glob($this->database."/".$this->schema."/".$mask.$this->docExt);
+        $list = glob($this->database."/".$this->collection."/".$mask.$this->docExt);
         foreach ($list as &$file) {
             $file=basename($file,$this->docExt);
         }
@@ -258,6 +318,7 @@ class DocumentStoreOne {
      * @return bool
      */
     private function lockFolder($filepath,$maxRetry=-1){
+        if ($this->manualLock!=null) return true; //it's already locked manually.
         clearstatcache();
         $maxRetry=($maxRetry==-1)?$this->defaultNumRetry:$maxRetry;
         $lockname=$filepath.".lock";
@@ -279,9 +340,11 @@ class DocumentStoreOne {
     /**
      * Unlocks a document
      * @param $filepath
+     * @param bool $forced
      * @return bool
      */
-    private function unlockFolder($filepath){
+    private function unlockFolder($filepath,$forced=false){
+        if ($this->manualLock!=null && !$forced) return true; // it's locked manually it must be unlocked manually.
         $unlockname= $filepath.".lock";
         return @rmdir($unlockname);
     }
